chore: refactor references updated
